🛠️refactor: 스터디룸 데이터 관리 로직 분리 및 훅, 라우트 폴더 구조 개선
1. 스터디룸 데이터 관리 로직 분리
- useStudyRooms 커스텀 훅 생성
- StudySection 컴포넌트에서 데이터 로직 분리

2. 훅 폴더 구조 개선
- src/hooks 하위에 기능별 폴더 구조화 (studyroom, article, ui 등)
- 각 폴더별 index.ts 파일로 export 관리

3. 라우트 폴더 구조 개선
- 불필요한 (private)과 (public) 라우트 그룹 삭제
