+++ src/app/_component/domain/login/NameBtn.tsx
@@ -7,8 +7,11 @@ import styles from "./NameBtn.module.css";
 import LogoutModalContent from "./modal/LogoutModalContent";
 
 export default function NameBtn() {
-  const { name, year } = useUserStore();
+  const name = useUserStore(state => state.name);
+  const year = useUserStore(state => state.year);
+
   const open = useModalStore(state => state.open);
+
   const handleOpenLogout = () => {
     open(<LogoutModalContent />);
   };
