+++ src/hooks/common/useConfirmBeforeRefresh.ts
