+++ src/hooks/microLink/useContentExtract.ts
